[BOO] Add a cache prepopulator (#642)
Adds a python class `CachePrepopulator` which can be used to prepopulate
the launchable cache, and a script for running the cache prepopulator
for a text file of miopen commands.

From python, a user could specify any combination of `torch.device`'s,
`ConvSignature`s, miopen driver commands, and a commands file path when
initializing `CachePopulator`.

Running the `CachePopulator` instance with the `run` method will
populate the `BOO_CACHE_DIR` with `.mlir` files and `.vmfb` files for
each signature and driver command. In subsequent python sessions, the
user can create a launchable from one of these conv signatures or
commands and it will automatically fetch the pre-compiled kernels when
running.

This PR also adds a few small changes:

1. Changes the function names to disambiguate layouts
2. Enables setting the global `CACHE_BASE_DIR` with a python function.
This was done to allow changing the cache dir for testing without
setting the environment variable.
3. Removes a completed `TODO` comment.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>