[Wave][Cache] Add caching support for nested functions
Currently, we are caching nested function as freevars which is detected
in getclosurevars. However we are only caching it's function pointer
which means every time we have a nested function inside wave kernel,
it will recompile every single time since it has a different function
pointer.

Even worse, the kernel defintion that we have cached for these nested
functions are often just something like:
```
    @tkw.wave(constraints)
    def base_attention(
        q: tkl.Memory[B, N_Q, D_Q, GLOBAL_ADDRESS_SPACE, LOGIT_DTYPE],
        k: tkl.Memory[B, N_KV, D_Q, ADDRESS_SPACE, LOGIT_DTYPE],
        v: tkl.Memory[B, N_KV, D_KV, ADDRESS_SPACE, LOGIT_DTYPE],
        c: tkl.Memory[B, N_Q, D_KV, GLOBAL_ADDRESS_SPACE, tkl.f32],
    ):
        base_attention_core(q, k, v, c)
```

Which is very incomplete since we do not know what is inside
base_attention_core and could change between different functions.

To solve this, we added a simple BFS search that detects all
subfunctions, and then we just accumulate all these function's generated
key to compute the final key. This way we can solve/cache all sorts of nested
function.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>