Inject cache directory into the cache manager (#590)
The current setup hard-codes the cache directory for the global cache
manager based on an environment variable. This is a problem particularly
in tests because it causes the tests to be non-hermetic and to mess with
the users cache directory.

I think it would be better if the tests could instantiate their own
cache manager explicity rather than relying on the global, but that
would require a more significant restructuring.

Note that this removes the lazy loading of cached kernel asm. I realized
that https://github.com/iree-org/iree-turbine/pull/574 effectively made
this useless as the only place where a `WaveCache` was being used, the
asm was immediately retrieved. If we want to restore that behavior then
we'd need to go back to returning a more complicated structure instead
of the asm, but it should be a consistent structure as discussed in
https://github.com/iree-org/iree-turbine/pull/574.

It incidentally removes a bunch of usages of the custom `_read_file` and
`_write_file` convenience functions that were being called on
`pathlib.Path` objects. These functions are pretty much redundant with
the `pathlib` standard library functions. I didn't go so far as purging
their usage in the rest of the codebase, as I didn't want to expand this
PR beyond the kernel cache, but I think they should probably be removed.

Fixes https://github.com/iree-org/iree-turbine/issues/420