Feat: better scaling for fp8 quant (#679)
Currently, there is a line that shifts the tensors before softmax
quantization, however that is effectively a no-op.

With this change, the value quantized to FP8 are correctly rescaled from
the range [0, 1] to [0, 240] before quantization, thus using more the
fp8 value range.

We need to undo the rescaling at the end before the output is returned.

cc @raikonenfnu @KyleHerndon

---------

Signed-off-by: Giuseppe <giuseppefranco4@gmail.com>