[Wave] Initial support for Multibuffering (#633)
This adds initial support for performing multibuffering on loads during
the scheduling guarded by the `SchedulingType` option.

This also adds a custom sympy operator for `xor` to model the switching
between the upper and lower part of a buffer during the loop.
The operator maps to `arith.xori` during codegen.

This feature is currently experimental, and in the future, we aim to
provide finer-grained control over which loads / stores to multi buffer.