[BOO] Adds generic convolution custom ops for BOO convs (#701)
This includes two `CustomOp`:

- `generic_conv : (Tensor x, Tensor w, int[] stride, int[] dilation, str
xl, str wl, str ol, int[] os) -> Tensor`, where `xl`, `wl`, `ol` are
string attributes controlling the layouts of `x`, `w`, and the output
`o`. These modify the indexing maps in the linalg generics, `os` is the
`output_shape` can be `[]` or an explicit list of ints to use for the
output shape. It is useful to be able to pass this attribute when the
output shape of the conv needs to be enlarged (bwd) or shrunk (wrw).
- `insert_slice : (Tensor src, Tensor dst, int[] offset, int[] stride)
-> Tensor` directly writes a `tensor.insert_slice` op into the IR. This
is to avoid unfortunate IR generated by pytorch code like `dst[indices]
= src`.

In addition:

- `ConvSignature` `get_nn_module(use_custom=True)` now leverages these
custom ops to produce transpose-less IR, and supports all grouped
backward configurations (`transposed=True` not yet implemented).

Mostly, this is a performance improvement with a few exceptions. Issues
have been posted in IREE for the most significant regressions. E.g.,
<https://github.com/iree-org/iree/issues/20498>,
<https://github.com/iree-org/iree/issues/20511>,
<https://github.com/iree-org/iree/issues/20473>.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>