[Wave] Rename reduction to iterate
This PR renames the tkw.reduction op to tkw.iterate.
While the original op had reduction semantics, with expansion
handling the unrolling of the reductions, this op can now
be expanded to an iteration over a dimension D.

This PR just renames the op. Many functions will still
refer to the variable as a reduction, but that modification
requires more careful thought.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>