[TKW] Use uptream `gpu.shuffle` repacking (#740)
AMDGPU shuffles only support i32 internally and all other datatypes need
to be bitcasted/zexted/repacked to multiple values. Instead of doing
this on Wave side (which was incomplete anyway), rely on upstream
repacking during GPU->ROCDL lowering, which was recently added.

This requires latest IREE so it's a breaking change.

Signed-off-by: Ivan Butygin <ivan.butygin@gmail.com>