[Wave] Add support for init args in while loops (#755)
This PR adds support for init args in the while loop style iterate ops.
In order to accomplish this, a couple of changes had to be made
1. Moving add_get_results out from an expansion specific pass to one
before the infer types pass
2. Fixing indices of MMA and ACC nodes during index propagation
3. Updates to handler. Specifically since the while loop returns the
value of the induction variable while we don't expose this in the DSL,
we reorder the arguments.
4. Changes to lit tests to fix this
5. Updates to logic to determine MMA reduction dims and added validation
to avoid ambiguities

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>