[Wave] Add wave_sdpa API for f16 attention (#773)
This PR adds a wave_sdpa API that is similar to the Pytorch API.
Internally, it utilizes the vanilla attention template to get the
kernel.

The template was modified to do the scaling inside the kernel so that
the query matrix no longer has to be scaled outside the kernel. This was
modified in all the users of the vanilla template.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>