[Wave] Generalize Barriers to work across iterations
This PR has two main purpose.

First and most importantly, it teaches our
compiler to only insert shared memory barriers when there is truly no
barriers in between ops. This was assumed before but with the
soon-coming PR of ping-pong where we re-order and want barriers to be
precisely inserted, this PR help handles those case and to not
accidentally insert too many barriers when not needed.

Secondly, and complimentarily, this PR teaches compiler to insert
barrier between ops from different iterations. Most commonly,
shared_read at time=t, and shared_write at time=t+1. This allows us to
also remove the hardcoded insertion of barrier at the start of every
Iterate subgraph. For pingpong, this will allow the PingPong pass to
manually insert barriers for next iteration, and this pass can analyze
properly to not introduce redundant barriers.

To achieve those two goals we set above, below are listed the changes in
code and their motivation:

- Fix bugs in decompose_reduce_ops where subgraph of Conditional is not
  properly registered to subgraphs and parent_op. Changes can be seen in
  tracing.py and decompose_reduce_ops.py

- Fix bug in pipelined_reduction where minimize_alloc cannot handle
  pipelined_reduce because it did not have a proper parent_op.
  Previously we have a workaround in tracing.py which is now removed.
  Changes can be seen in loop_reconstruction.py and tracing.py.

- Generalize insertion of barrier at top of Iterate into something that
  checks if barrier is required between ops from different iterations.
  This is done by adding support in `is_barrier_between` for case where
  the src do not dominate dst. We handle this by doing a check from src
  to end of loop, and then from start of loop to dst.

-  We also made changes into `barriers.py` to do one more round of
  `add_shared_memory_barriers`, after the first round to mimic the next
   step of iteration, and see if there are any inter-iteration dependencies.
   This naturally removes the hardcoded version of inserting barrier at
   top of loop.

- Many LIT tests are updated because of the two points above, instead of
  having lds_barrier at top of every iterate, we now have barrier
  inserted only right before it's actually used which is the
  first shared_write in a loop (protecting shared read of previous iter).

Signed-off-by: Stanley <stanley.winata@amd.com>