Moves `no_sync` to a `Device` property `sync` and stores this as an attribute for the device interop
Although it might not be particularly useful right now, this bool is used to enable the old synchronization interop.
E.g. if someone was launching kernels on a `Device` instantiated from a `uri` independent from pytorch.

Most importantly, the cache keys for `_CURRENT_THREAD.device_by_torch_device` are modified to include the current torch stream.
This way, a new `Device` will be created when we are launching kernels within a different torch stream.

Lastly, the `torch_stream` int is included in the `Device.instance_cache_key` so as to disambiguate between instances for different streams.

Signed-off-by: zjgarvey <zjgarvey@gmail.com>