[Wave] fix dim scaling to take into account "scaled" shape
Previously if shared memory is turned on global_read and local_write
gets expanded ~32 times(because dim scaling was wrong). So now we fixed
it to not expand that many time. It was not an issue with only global
because unlike shared_address_space with local_write, global does not
have write and hence is not a leaf_node which gets expanded by
dim_scaling.

Also modified minimiz_global_loads to have the same fix to work

Signed-off-by: Stanley Winata <stanley.winata@amd.com>