Reduce CPU overhead in `lookup_device_from_torch` (#905)
Calling `torch.cuda.current_stream(torch_device)` was taking between 15
and 50 us on the CPU for every launchable call.

This is because it checks whether device is an `int`, or a
`torch.device`, or a `str`, etc., lazy initializing cuda, and creating a
`torch.cuda.Stream` just so we could get the raw stream pointer int.

This PR also fixes some tests for Device that were not running due to
duplicate class name, and adds a few more checks on the GPU device
tests.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>