[BOO] Add some boo_driver improvements (#910)
1. Looks for multiple available GPUs and distributes the work across all
of them. I'm not sure if this is always desirable, so we could add
another command-line arg for toggling this.
2. Synchronizes all devices and collects garbage at reasonable points. I
chose to clean up after generating 96GB worth of output tensors. Some of
these outputs may already get cleaned up, but half the memory available
for an Mi300x (192 GB) seems like a very safe threshold to collect at.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>