Initial plumbing for Wave kernel integration with sharktank (#787)
This is the initial plumbing for integrating Wave kernels into
sharktank. The [wave
IR](https://gist.github.com/aviator19941/062c502e1b05f2eadcf514feb8a76413)
contains a func.func that calls a function in the stream.executable and
this PR merges the wave IR and creates a target_function that calls the
wave IR stream executable.

The Wave integration into sharktank makes use of the `mlir_kernel`
decorator in sharktank by passing in the MLIR from `wave_compile` along
with a `util.func` to call the FuncOp in the Wave asm. We need to add an
`inline` pass
[here](https://github.com/iree-org/iree-turbine/pull/787/files#diff-56777d4d58cb01861a15563fbccba4e55df2c83db374b5f3bca3c43d564a4097R143)
in order to remove the private `func.func` and `util.func` calls in the
MLIR to avoid having a `func.func @wave_kernel` that has the same body
of the `func.func @main` function. The inline pass removes the duplicate
`func.func @wave_kernel` and keeps the `func.func @main`.

---------

Signed-off-by: aviator19941 <avinash.sharma@amd.com>