[Wave] fix comparisons with `None`
In Python, `None` comparisons are tricky.  The `==` operator checks for
equivalence and it can return True under specific conditions like with
the `__eq__` function is defined for the class whose objects are
compared.  Instead, the reliable way to check for `None` is using the
`is` operator, which checks for identity instead of equivalence.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>