[boo] swap the import dependency condition for tracy
The original implementation for conv was importing the conv op
definition only when timing/tracing is enabled to avoid option clash for
IREE runtime that is unconditionally loaded when traversing the python
package tree for iree.turbine.kernel. This made it virtually impossible
to refactor out the tracing-related code.

Swap the dependency condition. This requires the traced function to be
placed in a fully separated process, not a fork of the current process.
This has a positive side effet of being runnable on all operating
systems and not only those supporting process forks. It also removes the
need for the user to explicitly mess the environment when they want
tracing as we can set up environment variables programmatically for
spawed processes.

Signed-off-by: Alex Zinenko <git@ozinenko.com>