[BOO] Add Graph-Based Fusion API (#967)
Adds some tooling for replacing subgraphs with auto-generated iree
launchable calls compatible with autograd and torch.compile.

### Approach

1. Identify root ops we want to offload to iree.
2. Indicate good consumer/producer fusion opportunities for each root
op. (See `FusionSchema` and `OpFusionSpec`).
3. Given a `torch.nn.Module`, use `torch.export.export` to get an
`ExportedProgram` (contains a `GraphModule`).
4. Extract all subgraphs consisting of root op fusions.
5. Replace each subgraph with a single node representing an iree-turbine
`Launchable` wrapped in an `autograd.Function`.
6. Return the module associated with the `ExportedProgram`, which
applies the modified graph.

### Some considerations

This is currently aot-only, as we are swapping subgraphs in an
aot-traced dynamo export. We could add support for dynamic shapes in the
future, or try to build something which will jit export a graph and do
the replacement on the fly. However, I feel like those use-cases might
be better served by using our `torch.compile` backend selectively.

The user currently needs to provide a `FusionSchema`, but we should
provide defaults for supported fusions.

The auto-generated graph ops in `fusion_transform` do not explicitly
support automatic mixed-precision contexts, nor do they have
custom-logic for alternative memory formats such as
`torch.channels_last`. In order to make this work compatible with
`boo_convolution`, which does support alternative memory formats, we
need to handle this properly. We should also make `boo_convolution` a
turbine `CustomOp` so we can export graphs with `boo_convolution` ops.

We are not attaching the `make_single_dispatch` attribute to the
generated mlir. It might be good to enable this for forward fusion
graphs by default.

### Some critical issues

1. There may be an issue with subgraph replacement when multiple
subgraphs match one pattern (but the first of the two subgraphs had
inputs which did not require gradient). It might be a good idea to do
something in-house for subgraph replacement, since we know exactly where
we want to put the replacements. Alternatively, we could make the
backward logic a bit more careful, but I'm not sure how to make this
work yet.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>