Disable broken CI runner
This runner has been down for over 2 weeks with a system error (see https://github.com/iree-org/iree-turbine/actions/runs/16127205662/job/45506873739?pr=1012#step:2:33 for example), leading to all PRs landing in an overall "failing". This is dangerous as it creates a habit of ignoring CI failures and ultimately defies the purpose of having CI. Disabling this runner until it can be fixed.