[boo] refactor cache prepopulator towards generalization
Move the cache prepopulator to the generic driver directory and
generalize it based on the previously introduce OpSignature and
OpCLIParser classes in preparation for adding more ops.

Add/refine type annotations and fix typos while I'm here.

The shim in `conv_exports` is preserved because we don't have a parser
dispatcher yet.

Signed-off-by: Alex Zinenko <git@ozinenko.com>