Workgroup Reordering Constraint + Template (#1000)
This is code that I have written to allow end-users to implement
workgroup reordering. I added a new type of Constraint called
ReorderingConstraint. At the moment, it's a very basic class, and it
doesn't extend the base Constraint class either (not sure if this is
needed?). I also created a template for a gemm kernel that includes this
ReorderingConstraint as well as a separate test file to test the kernel.

I will add my documentation (rst file) soon as well.

Signed-off-by: bodhisaha <bodhisaha04@gmail.com>