[Wave] Build wave runtime on demand
This PR adds the ability to build the wave runtime the first
time someone attempts to import it. This is done as a stepping
stone to enable wave runtime to work with iree-turbine
pypi installs.

We accomplish this by packaging setup.py, CMakeLists.txt and
runtime.cpp with the pypi package and then do a pip install
on import. When a user attempts to upgrade to a new
version of iree-turbine, they will have to manually
pip uninstall wave-runtime.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>