Optional tensor arguments for `CustomOp`. (#1044)
This PR implements optional arguments for `CustomOp`. The intended
consumption pattern is a `CustomOp`s `select` method chooses a variant
based on the presence of optional argument read using `KernelSelection`
then in the `generate` method produces IR to handle the arguments
already passed.

This PR changes the (I think intended) semantics of `None` in
`KernelSelection` somewhat. Some assertions in the code indicated that
`None` would be a valid `ArgumentDescriptor` for optional values.
Instead, this implementation asserts that `None` always indicates an
uninitialized argument (not read from actual values at the call site).
Note that `None` meant an uninitialized value before, now it just
*always* means that. Any optional values that weren't provided are
intended to be marked with sentinel types that indicate "we looked for
this optional value, but one wasn't passed".

---------

Co-authored-by: zjgarvey <47986913+zjgarvey@users.noreply.github.com>