[Wave] Add outer loop to search for schedules (#1019)
This PR adds code to search over schedules for a given kernel. In this
particular case, we choose attention and tweak the schedule of nodes in
the graph to tweak performance. We also clean up the print utils to
enable some of these changes. Finally, a new section has been added to
the documentation describing the code and how it can be used.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>