[boo] introduce op registry and refactor op exports
Introduce a registry class handling metadata classes for BOO ops. This
automatically traverses the ops available in `op_exports` and registers
them with a singleton registry class that can be later queried using the
op "name". This name is supposed to be a unique string that cannot
appear in the CLI syntax of another op. This is suboptimal and is
difficult to ensure statically without a formal syntax, but will work as
long as the syntax mirrors that of MIOpen driver.

This allows a major refactoring of op exports that can now be aggregated
in a single directory and share generic implementation for benchmarking,
numerics, and cache prepopulation. Note that most of the implementation
was refactored previously leaing only shims in specific op exports.

Most of this commit is moving code around. Cache pre-population test is
updated to exercise the multi-op parser and dispatch. Driver files are
updated to use the registry instead of directly loading individual ops.
READMEs are consolidated into the top-level user-facing document and
developer-oriented document in the `driver` directory.

Signed-off-by: Alex Zinenko <git@ozinenko.com>