[BOO] Enable more generic layout handling (#1086)
In anticipation of supporting `torch.nn.Linear` through boo graph
fusion, we need to enable layout handling for things like simple
transposes.

For example, given the graph for a `torch.nn.Linear` module:

```
class GraphModule(torch.nn.Module):
    def forward(self, primals_1: "f32[16, 64][64, 1]", primals_2: "f32[16][1]", primals_3: "f32[32, 64][64, 1]"):
         # File: /home/zjgar/code/iree-turbine/turbine.venv/lib/python3.12/site-packages/torch/_dynamo/external_utils.py:70 in inner, code: return fn(*args, **kwargs)
        t: "f32[64, 16][1, 64]" = torch.ops.aten.t.default(primals_1);  primals_1 = None
        addmm: "f32[32, 16][16, 1]" = torch.ops.aten.addmm.default(primals_2, primals_3, t);  primals_2 = t = None
        return (addmm, primals_3)
```

The input tensor `t` to `addmm` is not contiguous, nor is it in
`channels_last` format. If we are simply exporting the `addmm` op, then
this PR will transpose `t` back to contiguous format before `dlpack`
handoff to IREE, and add in the inverse transpose into the exported
MLIR. This will avoid making any mem copies, as the pytorch transpose
will simply modify tensor metadata, and the transpose in IREE should
fuse with the matmul.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>