[boo] introduce op registry (#1078)
Introduce a registry class handling metadata classes for BOO ops. This
automatically traverses the ops available in `op_exports` and registers
them with a singleton registry class that can be later queried using the
op "name". This name is supposed to be a unique string that cannot
appear in the CLI syntax of another op. This is suboptimal and is
difficult to ensure statically without a formal syntax, but will work as
long as the syntax mirrors that of MIOpen driver.

---------

Signed-off-by: Alex Zinenko <git@ozinenko.com>