Support for reading in binary files back into torch format (#1068)
`write_raw_tensor` is useful for passing tensors of dtypes not supported
by numpy. For testing reasons in shark-ai we want to be able to save
that file and invoke iree-run-module or iree-benchmark-module with it as
well as read it back from file, to make sure we're working with the
exact same input, and run other tests.

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>