[boo] Convert tests to `pytest` (#1100)
I'd like to introduce pytest fixtures for simplifying the
setup/resetting of various boo settings (e.g. caching). Converting our
`unittest`-style tests is the first step towards that.

Some of the test file structure is unnecessary now (e.g. test classes),
but I've left them as-is to minimize the diff. I've also confirmed we
have the same number of tests (149) before and after this change.