[boo] Add a pytest fixture for enabling the file cache (#1101)
This adds a `boo_cache_dir` pytest fixture, which enables the file cache
in a test-local directory for the duration of the test, and returns the
path so that the test can inspect the contents. This avoids each test
needing to set up its own temporary directory. A following change will
turn off the file cache by default, so only tests that explicitly
request `boo_cache_dir` (or turn on the cache internally) will have file
caching enabled.

A change in behaviour here is that the cache is left intact after test
execution finishes, which allows for easier debugging. By default pytest
dumps test files under `/tmp/pytest-of-$USER/pytest-current/`

This also adds an `autouse` fixture which sets the BOO cache directory
to a test-local directory, in order to isolate tests from each other.