[device] Use SKU as HIP target when possible (#1111)
Supplying a specific SKU will allow IREE to optimize for specific target
details.

This is implemented as a table mapping torch device names to IREE SKU
strings. For now I've included the known Instinct devices listed here:

https://github.com/iree-org/iree/blob/66b2f337c8019abf4716487ad32adea032afbb81/compiler/src/iree/compiler/Codegen/Dialect/GPU/TargetUtils/KnownTargets.cpp#L425-L440

Unfortunately this requires keeping the table in sync with IREE as new
targets are added.