[BOO] Don't use `TensorMetadata` to infer output layouts (#1121)
When a conv output has all unit spatial dims, it is in both contiguous
format and channels last format. There was an inconsistency between the
generation of the graph op inferring the memory format from the
`TensorMetadata`, and the conv replacement tool inferring the layout
from the fake tensor.

Unfortunately, this change can't easily be tested, since pytorch/miopen
seems to unreliably select whether to return channels last or contiguous
outputs for conv.

Signed-off-by: zjgarvey <zjgarvey@gmail.com>